component model refactoring (#221)
