+++ agent/agentUi/src/NewAccessModal.jsx
@@ -1,0 +1,52 @@
+import {Box, Button, Modal, TextField} from "@mui/material";
+import {useFormik} from "formik";
+import {modalStyle} from "./model/theme.js";
+
+const NewAccessModal = (props) => {
+    const newAccessForm = useFormik({
+        initialValues: {
+            token: "",
+            bindAddress: "",
+        },
+        onSubmit: (v) => {
+            props.handler(v);
+            props.close();
+        },
+    });
+
+    return (
+        <Modal
+            open={props.show}
+            onClose={props.close}
+        >
+            <Box sx={{ ...modalStyle }}>
+                <h2>Access...</h2>
+                <form onSubmit={newAccessForm.handleSubmit}>
+                    <TextField
+                        fullWidth
+                        id="token"
+                        name="token"
+                        label="Share Token"
+                        value={newAccessForm.values.token}
+                        onChange={newAccessForm.handleChange}
+                        onBlur={newAccessForm.handleBlur}
+                        sx={{ mt: 2 }}
+                    />
+                    <TextField
+                        fullWidth
+                        id="bindAddress"
+                        name="bindAddress"
+                        label="Bind Address"
+                        value={newAccessForm.values.bindAddress}
+                        onChange={newAccessForm.handleChange}
+                        onBlur={newAccessForm.handleBlur}
+                        sx={{ mt: 2 }}
+                    />
+                    <Button color="primary" variant="contained" type="submit" sx={{ mt: 2 }}>Create Access</Button>
+                </form>
+            </Box>
+        </Modal>
+    );
+}
+
+export default NewAccessModal;
