+++ agent/agentUi/src/model/theme.js
@@ -1,0 +1,46 @@
+const componentOptions = {
+    MuiCard: {
+        styleOverrides: {
+            root: ({theme}) => theme.unstable_sx({
+                mt: 5,
+                p: 2.5,
+                pt: 3.25,
+                borderRadius: 2,
+            }),
+        }
+    },
+    MuiAppBar: {
+        styleOverrides: {
+            root : ({theme}) => theme.unstable_sx({
+                borderRadius: 2,
+            }),
+        }
+    }
+}
+
+export const themeOptions = {
+    components: componentOptions,
+    palette: {
+        mode: 'light',
+        primary: {
+            main: '#241775',
+        },
+        secondary: {
+            main: '#9bf316',
+        },
+    },
+    typography: {
+        fontFamily: 'Poppins',
+    }
+};
+
+export const modalStyle = {
+    position: 'absolute',
+    top: '25%',
+    left: '50%',
+    transform: 'translate(-50%, -50%)',
+    width: 600,
+    bgcolor: 'background.paper',
+    boxShadow: 24,
+    p: 4,
+};
